+++ protocol/clproto_cpp/test/tests/test_joint_space_proto.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include <state_representation/robot/Jacobian.hpp>
+#include <state_representation/space/Jacobian.hpp>
 #include <state_representation/robot/JointState.hpp>
 #include <state_representation/robot/JointPositions.hpp>
 #include <state_representation/robot/JointVelocities.hpp>
