+++ protocol/clproto_cpp/test/tests/test_packing.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 
 #include <state_representation/space/cartesian/CartesianState.hpp>
-#include <state_representation/robot/Jacobian.hpp>
+#include <state_representation/space/Jacobian.hpp>
 #include <state_representation/robot/JointState.hpp>
 
 #include "clproto.h"
