+++ source/state_representation/test/tests/robot/test_jacobian.cpp
@@ -1,4 +1,4 @@
-#include "state_representation/robot/Jacobian.hpp"
+#include "state_representation/space/Jacobian.hpp"
 #include "state_representation/exceptions/IncompatibleStatesException.hpp"
 #include <gtest/gtest.h>
